Refactor: MemberQueryResponseDTO - 중첩 클래스 분리
