+++ src/main/java/ticketaka/mtvs3_final_backend/sticker/query/repository/StickerQueryRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ticketaka.mtvs3_final_backend.file.command.domain.model.property.RelationType;
 import ticketaka.mtvs3_final_backend.member.query.dto.MemberQueryResponseDTO;
+import ticketaka.mtvs3_final_backend.member.query.dto.getMemberStickerDTO;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.Sticker;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.StickerRarity;
 import ticketaka.mtvs3_final_backend.sticker.command.domain.model.StickerType;
@@ -26,13 +27,13 @@ public interface StickerQueryRepository extends JpaRepository<Sticker, Long> {
 
     List<Sticker> findAllByConcertIdAndStickerTypeAndStickerRarity(Long concertId, StickerType stickerType, StickerRarity stickerRarity);
 
-    @Query("SELECT new ticketaka.mtvs3_final_backend.member.query.dto.MemberQueryResponseDTO.getMemberStickerDTO(" +
+    @Query("SELECT new ticketaka.mtvs3_final_backend.member.query.dto.getMemberStickerDTO(" +
             "s.id, s.stickerName, s.stickerScript, s.stickerRarity, f.fileUrl) " +
             "FROM Sticker s " +
             "JOIN MemberSticker ms ON s.id = ms.stickerId " +
-            "LEFT JOIN File f ON f.relationType = :relationType AND f.relationId = :relationId " +
+            "LEFT JOIN File f ON f.relationType = :relationType AND f.relationId = s.id " +
             "WHERE ms.memberId = :memberId AND s.stickerType = :stickerType")
-    List<MemberQueryResponseDTO.getMemberStickerDTO> findAllByMemberIdAndStickerTypeAndRelationType(@Param("memberId") Long memberId,
-                                                                                                    @Param("stickerType") StickerType stickerType,
-                                                                                                    @Param("relationType") RelationType relationType);
+    List<getMemberStickerDTO> findAllByMemberIdAndStickerTypeAndRelationType(@Param("memberId") Long memberId,
+                                                                             @Param("stickerType") StickerType stickerType,
+                                                                             @Param("relationType") RelationType relationType);
 }
